Different implementation, easier to test
