+++ lib/tomo/plugin/core.rb
@@ -5,16 +5,6 @@ module Tomo::Plugin
   module Core
     extend Tomo::PluginDSL
 
-    class << self
-      private
-
-      def local_user
-        ENV["USER"] || ENV["USERNAME"] || `whoami`.chomp
-      rescue StandardError
-        nil
-      end
-    end
-
     helpers Tomo::Plugin::Core::Helpers
     tasks Tomo::Plugin::Core::Tasks
 
@@ -26,13 +16,13 @@ module Tomo::Plugin
       keep_releases:     10,
       linked_dirs:       [],
       linked_files:      [],
-      local_user:        local_user,
+      local_user:        nil, # determined at runtime
       release_json_path: "%<release_path>/.tomo_release.json",
       releases_path:     "%<deploy_to>/releases",
       revision_log_path: "%<deploy_to>/revisions.log",
       shared_path:       "%<deploy_to>/shared",
       tmp_path:          "/tmp/tomo",
-      run_args:          []
+      run_args:          [] # determined at runtime
     )
   end
 end
