+++ lib/tomo/testing/plugin_tester.rb
@@ -25,10 +25,6 @@ module Tomo
         end
       end
 
-      def settings
-        runtime.execution_plan_for([]).settings
-      end
-
       private
 
       attr_reader :host, :runtime
